feat(pipeline): allow the build step to have its own runner configuration (#10)
This change allows the `build` (synth) step to have its own Github
Runner configuration that is separate from the deployment steps. This
allows the build step to have faster/larger hardware without impacting
the cost of the other steps. I've also added in the ability to pass in
runner groups.