feat(stages): allow providing per-stage aws credentials (#15)
The existing code made the assumption that the entire `GithubPipeline`
would operate with the same AWS credentials throughout. That's not
always going to be the case. It's reasonable that you may have a single
`GithubPipeline` that first deploys to one account, and then deploys to
a second account.

This PR introduces an override at the `GithubStage` level that allows
overriding the `AwsCredentialsProvider` for any stacks within that
`GithubStage`.

**Note on the `Wave`**

Currently I am leaving the `Wave` out of it and purely doing this work
in the `GithubStage`. After we test it a bit, we should be able to add
this pattern into the `GithubWave` as well.

Co-authored-by: Matt Wise <diranged@users.noreply.github.com>